requirements.txt
faiss-cpu==1.7.4
numpy
requests
sentence_transformers==2.2.2
openai